+++ src/main/java/com/acme/banking/dbo/ooad/dao/AccountRepository.java
@@ -1,4 +1,6 @@
-package com.acme.banking.dbo.ooad.domain;
+package com.acme.banking.dbo.ooad.dao;
+
+import com.acme.banking.dbo.ooad.domain.Account;
 
 import java.util.Collection;
